machine-learning/Sentiment/requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud-language>=2.5.1

